+++ src/pages/Search/Searching/SearchResults/SearchResults.test.tsx
@@ -1,10 +1,13 @@
 import { render } from '@testing-library/react';
 import React from 'react';
 import { BrowserRouter as Router } from 'react-router-dom';
-import { SearchStatus } from '../useSearchPageFacade';
-import { CommPkgPreview, CompletionStatus } from '../../../services/apiTypes';
+import { SearchStatus } from '../../useSearchPageFacade';
+import {
+    CommPkgPreview,
+    CompletionStatus,
+} from '../../../../services/apiTypes';
 import SearchPage from './SearchResults';
-import { testCommPkgPreview } from '../../../test/dummyData';
+import { testCommPkgPreview } from '../../../../test/dummyData';
 
 describe('<SearchResult/>', () => {
     it('Renders placeholder text before user starts typing', () => {
